+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+
+- Fix unwanted schema validation (#392)
+
 ## v1.18.0 - (2022-07-19)
 
 - Overwrites - Automatically convert values for request query parameters, path variables & headers to string (#336, #384)